Finalized the first version of the GUI scraper; it still needs some fine tuning for updates (it currently parses all months since February, 2006 every time it runs)

+++ BusinessEntities/DBEntities/VoteDetailDBE.cs
@@ -8,7 +8,7 @@ namespace ro.stancescu.CDep.BusinessEntities
 {
     public class VoteDetailDBE
     {
-        public virtual int? Id { get; set; }
+        public virtual UInt64? Id { get; set; }
 
         public virtual VoteSummaryDBE Vote { get; set; }
